Update chip field to make it required
We might need to make upserts of pets, which means that we need to have an unique identifier for each pet. Pets don't have email addresses, so we will use the chip for that. It is not exactly matching the reality of pet identification, but it will suffice for our app